requirements.txt
certifi==2018.10.15
chardet==3.0.4
click==6.7
fancycompleter==0.8
Flask==0.12.2
Flask-Cache==0.13.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==1.2.6
pdbpp==0.9.2
Pygments==2.2.0
pyrepl==0.8.4
python-dateutil==2.7.3
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
Werkzeug==0.12.2
wmctrl==0.3
